requirements.txt
aiohttp
aiosignal
async-timeout
attrs
backoff
certifi
charset-normalizer
frozenlist
greenlet
idna
multidict
numpy
pandas
python-dateutil
python-dotenv
pytz
requests
six
SQLAlchemy
typing-extensions
urllib3
yarl
azure-storage-file-datalake
